Code more services
also some new unit tests and update some repositories

+++ ITAssetTracking.App/Services/AssetService.cs
@@ -14,7 +14,7 @@ public class AssetService : IAssetService
         _assetRepo = assetRepository;
     }
     
-    public Result<Asset> GetAssetById(int assetId)
+    public Result<Asset> GetAssetById(long assetId)
     {
         try
         {
