+++ ITAssetTracking.Core/Interfaces/Services/ISoftwareRequestService.cs
@@ -11,7 +11,7 @@ public interface ISoftwareRequestService
     Result<List<SoftwareAssetRequest>> GetOpenSoftwareRequests();
     Result<List<SoftwareAssetRequest>> GetRequestsByEmployee(int employeeId, bool includeClosed = false);
     Result<List<SoftwareAssetRequest>> GetRequestsBySoftware(int softwareAssetId, bool includeClosed = false);
-    Result<List<SoftwareAssetRequest>> GetRequestsByAsset(int assetId, bool includeClosed = false);
+    Result<List<SoftwareAssetRequest>> GetRequestsByAsset(long assetId, bool includeClosed = false);
     Result<List<SoftwareAssetRequest>> GetRequestsByResult(int resultId);
     Result<List<SoftwareAssetRequest>> GetRequestsInDateRange(DateTime startDate, DateTime endDate);
